requirements.txt
streamlit==1.55.0
langchain==0.3.28
langchain-core==1.2.18
langchain-community==0.3.24
langchain-text-splitters==0.3.11
langchain-google-genai==4.2.1
google-generativeai==0.8.6
PyPDF2==3.0.1
chromadb==1.5.5
python-dotenv==1.2.2